code cleaning and code testing
